IHttpModule: move to a single handler
This is a simpler approach that also allows async access to everything else (next step).
